Shortest-Network-Algoritms-Api/requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.31.0
networkx==3.3
